requirements.txt
streamlit==1.38.0
langchain==0.2.14
langchain-core==0.2.34
langchain-community==0.2.12
langchain-google-genai==1.0.7
google-generativeai==0.7.2
pydantic>=2.6,<3
faiss-cpu==1.8.0
tiktoken==0.7.0
pdfplumber==0.11.4
rank_bm25>=0.2.2